backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["github-actions <jojoto8845@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
flask-pymongo = "^2.3.0"
flask-cors = "^4.0.1"
mongoengine = "^0.28.2"
pymongo = "^4.8.0"
flake8 = "^7.1.0"
black = "^24.4.2"
isort = "^5.13.2"
mypy = "^1.10.1"
safety = "^3.2.4"
bandit = "^1.7.9"
google-auth = "^2.32.0"
openai = "^1.35.12"
python-dotenv = "^1.0.1"
bcrypt = "^4.1.3"

[tool.ruff]
target-version = "py311"
line-length = 119
exclude = [".mypy_cache"]
 
[tool.ruff.lint]
select = ["E", "W", "F", "I", "C90"]
ignore = ["E203"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
